Vfs rewrite (#21)
Rewrite the vfs. Before, to support a new file system, you would have to implement a whole bunch of structs and traits to map your file system from whatever it is to an Inode based structure. If the fs is already inode based, that meant a lot of 1:1 mapping.
Now, you implement the `FileSystem` trait, which makes extensive use of `VfsHandles`, that the separate `FileSystem`s can handle as they please.