make use of a Pin for the last stack pointer
During rescheduling, Threads are moved. Threads hold their last stack
pointer, which is required for multitasking. Until now, that field was
a simple usize, which could potentially be moved. The rescheduling logic
however relies on the memory location of that field.

With this fix, we use a Pin<Box<usize>> to guarantee that the field does
not move (specifically, that the pointer to that field that we use for
the task switch is valid when it's used).