implement unix sockets (#22)
* remove the vga stuff

* move init stuff into rt module in our std

Heap initialization is now done in the std crate.
This means that other userspace programs can now
just call `std::rt::init()` to get heap. From then
on, the program can use heap allocations.

* add a print and println macro to std

Userspace programs can now use println! and print!
which will just be printed to /dev/stdout.
Effectively, that text will then be written to the
serial port, but that will change in the future.

* remove obsolete comment

* no longer dump the process tree upon sys_write

* add window_server program

* move panic handler into std

* move non-arch specific stuff out of `arch`

* implement sys_socket stub

* implement sys_socket

For now, a socket is a unix socket. This changes
later to support more socket types.

* implement a sys_bind stub