add support for passing one argument during thread creation
This single argument has to be a *mut c_void unfortunately.
When using a generic <T>, we need another indirection that wraps the T in a Box, obtains that pointer, and on the other side of the FFI (namely when the thread gets scheduled the first time), obtain a Box<T> from that pointer again, then use Box::into_inner to obtain the T.
This indirection must have a stable ABI, needs to have a generic <T> and needs to have an accessible address that is known at compile time.
The signature `extern "C" fn trampoline::<T>(f: *extern "C" fn(), arg: *mut T)` does not work for some reason.
Godbolt understands the resulting assembly, but QEMU decides that a #GP is appropriate.

If you think this should work, please implement it, because that would be really nice to have. For now, you have to manually make sure that you unbox what you passed in.