fix process cleanup
Processes hold resources that are related to their address space, such as vm objects.
Previously, they were cleaned up in the kernel address space, which means that their virtual addresses were incorrectly interpreted.
With this change, process cleanup is performed on sys_exit.
Every process must call sys_exit to terminate.
If not, this will currently panic.

Instead of panicking, we could instead call sys_exit as syscall in the kernel (we need the change in privilege level, as the process might run in userspace).