make handle_packet more generic
By accepting anything that a packet can be made of (with TryFrom), this can later be re-used for udp, which computes its checksum over the IP packet.
With this design, the `TryFrom<OtherPacket>` impl can be seen as "how do I extract the relevant information from the parent packet/frame", while the `TryFrom<&[u8]>` impl is "how do I actually create the packet from the relevant data".