panic if the ethernet frame's payload is too large
On the one hand, it would be easy to return an error instead. On the other hand, if another protocol incorrectly segments their data, returning an error will probably not make them reduce the segment size, so we should just aim to implement that segmentation correctly in the first place.
It probably makes sense to revisit this decision in the future once we have more data link protocols with different MTUs.